# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriLance freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      summary: Admin - block or unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBlockUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/verify:
    patch:
      operationId: adminVerifyUser
      tags: [admin]
      summary: Admin - set verified/top-rated status on a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/role:
    patch:
      operationId: adminChangeUserRole
      tags: [admin]
      summary: Admin - change a user's role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminChangeUserRoleBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/jobs:
    get:
      operationId: adminListJobs
      tags: [admin]
      summary: Admin - list all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: flagged
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /admin/jobs/{id}/flag:
    patch:
      operationId: adminFlagJob
      tags: [admin]
      summary: Admin - flag or unflag a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFlagBody"
      responses:
        "200":
          description: Job flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /admin/audit-logs:
    get:
      operationId: getAdminAuditLogs
      tags: [admin]
      summary: Admin - fetch audit log entries
      parameters:
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: entityType
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
        status:
          type: string
        isFlagged:
          type: boolean
        proposalCount:
          type: integer
        clientName:
          type: ["string", "null"]
        clientAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientId
        - title
        - description
        - category
        - skills
        - budgetType
        - remote
        - status
        - isFlagged
        - proposalCount
        - createdAt
        - updatedAt
...
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
      required:
        - jobs
        - total
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isBlocked:
          type: boolean
        isVerified:
          type: boolean
        isTopRated:
          type: boolean
        verificationDocUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
        completedJobs:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - role
        - name
        - isBlocked
        - completedJobs
        - createdAt
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required:
        - users
        - total
    AdminBlockUserBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required:
        - isBlocked
    AdminVerifyUserBody:
      type: object
      properties:
        isVerified:
          type: boolean
        isTopRated:
          type: boolean
    AdminChangeUserRoleBody:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [freelancer, client, admin]
...
    AdminFlagBody:
      type: object
      properties:
        isFlagged:
          type: boolean
      required:
        - isFlagged
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        adminId:
          type: integer
        adminName:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        entityName:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - adminId
        - adminName
        - action
        - entityType
        - entityId
        - createdAt
    AuditLogListResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
      required:
        - logs
        - total